Switch to HashRouter
Deploying to GitHub Pages as a subfolder of my subdomain created the
issue of NOTHING APPEARING AT ALL in the <main> tag.
BrowserRouter takes the ENTIRE path and does not care that the app's
root folder is a SUBFOLDER. Oh well. Added a 404 page so I see this more
easily next time.